+++ src/commands.js
@@ -1,4 +1,5 @@
 import {
+  resetSeed,
   randomByte,
   randomShort,
   randomInt,
@@ -7,6 +8,8 @@ import {
   randomString,
   randomName,
   randomCity,
+  randomCountryCode,
+  randomCountryName,
   randomStreetAddress,
   randomPhoneNumber,
   randomEmail
@@ -22,6 +25,8 @@ export const extensionCommands = [
   { key: constants.CMD_KEY_RANDOM_GUID, callback: randomGuid },
   { key: constants.CMD_KEY_RANDOM_NAME, callback: randomName },
   { key: constants.CMD_KEY_RANDOM_CITY, callback: randomCity },
+  { key: constants.CMD_KEY_RANDOM_COUNTRY_CODE, callback: randomCountryCode },
+  { key: constants.CMD_KEY_RANDOM_COUNTRY_NAME, callback: randomCountryName },
   { key: constants.CMD_KEY_RANDOM_STREET_ADDRESS, callback: randomStreetAddress },
   { key: constants.CMD_KEY_RANDOM_PHONE_NUMBER, callback: randomPhoneNumber },
   { key: constants.CMD_KEY_RANDOM_EMAIL, callback: randomEmail }
@@ -34,5 +39,16 @@ export const extensionCommandsWithInput = [
     prompt: constants.MSG_ENTER_STRING_LENGTH,
     validation: isNumber,
     errorMsg: constants.MSG_INPUT_VALUE_MUST_BE_NUMBER
+  }
+]
+
+export const extensionExecCommands = [
+  {
+    key: constants.CMD_KEY_RESET_SEED,
+    callback: resetSeed,
+    prompt: constants.MSG_ENTER_SEED,
+    placeHolder: constants.MSG_LEAVE_EMPTY_FOR_DEFAULT,
+    errorMsg: constants.MSG_SEED_RESET_ERROR,
+    infoMsg: constants.MSG_SEED_RESET
   }
 ]