+++ src/constants.js
@@ -1,3 +1,4 @@
+export const CMD_KEY_RESET_SEED = 'extension.resetSeed'
 export const CMD_KEY_RANDOM_BYTE = 'extension.randomByte'
 export const CMD_KEY_RANDOM_SHORT = 'extension.randomShort'
 export const CMD_KEY_RANDOM_INT = 'extension.randomInt'
@@ -7,11 +8,17 @@ export const CMD_KEY_RANDOM_STRING_CUSTOM_LENGTH = 'extension.randomStringCustom
 export const CMD_KEY_RANDOM_NAME = 'extension.randomName'
 export const CMD_KEY_RANDOM_STREET_ADDRESS = 'extension.randomStreetAddress'
 export const CMD_KEY_RANDOM_CITY = 'extension.randomCity'
+export const CMD_KEY_RANDOM_COUNTRY_CODE = 'extension.randomCountryCode'
+export const CMD_KEY_RANDOM_COUNTRY_NAME = 'extension.randomCountryName'
 export const CMD_KEY_RANDOM_PHONE_NUMBER = 'extension.randomPhoneNumber'
 export const CMD_KEY_RANDOM_EMAIL = 'extension.randomEmail'
 
 export const MSG_NO_ACTIVE_TEXT_EDITOR = 'No active text editor found!'
 export const MSG_ENTER_STRING_LENGTH = 'Enter the length of the string'
 export const MSG_INPUT_VALUE_MUST_BE_NUMBER = 'The input value must be a number'
+export const MSG_ENTER_SEED = 'Enter new random seed'
+export const MSG_LEAVE_EMPTY_FOR_DEFAULT = 'Leave empty to use default options'
+export const MSG_SEED_RESET = 'Random seed was reset'
+export const MSG_SEED_RESET_ERROR = 'Invalid random seed'
 
 export const VALUE_DEFAULT_STRING_LENGTH = 10