+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to the "vscode-random" extension are documented in this file.
 
+## [1.1.0] - 2017-01-08
+## Added
+- New command to allow random seed reset
+- New command to return a random country code
+- New command to return a random country name
+
+## Changed
+- Code improvements
+
 ## [1.0.2] - 2017-01-04
 ### Fixed
 - Fix bug when generating random values with multiple editors